+++ variables.tf
@@ -1,13 +1,19 @@
 variable "chart_version" {
   description = "Helm chart version to install"
   type        = string
-  default     = "1.16.1"
+  default     = "1.29.0"
 }
 
 variable "cluster_ip" {
   description = "The cluster IP for the kube-dns service."
   type        = string
   default     = "172.20.0.10"
+}
+
+variable "k8s_app_label_override" {
+  description = "Override the k8s-app label for the CoreDNS deployment."
+  type        = string
+  default     = "kube-dns"
 }
 
 variable "replica_count" {