+++ CHANGELOG.md
@@ -15,11 +15,13 @@ At release time, move the entries from here to a new release section.
 ### Additions
 
 - Added `topology_spread_constraints` variable to allow distribution of replicas across availability zones.
+- Added `k8s_app_label_override` variable (defaults to `kube-dns`).
 
 ### Fixes
 
 ### Changes
 
+- Updated default `chart_version` to `1.29.0` to support `topologySpreadConstraints`.
 - Updated default `replica_count` to 3 for improved high availability.
 - Refactored `helm_release` to use `yamlencode` for all values.